+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v1.1.0] - 2025-07-04
 ### Fixed
 - [#43] Exclude ces-exporter from multinode system info 
 - [#45] checking dogu health when checking export-mode